requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.3.0
fastapi==0.121.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
pydantic_core==2.41.5
pydantic==2.12.4
Pygments==2.19.2
pymongo==4.15.4
PyMuPDF==1.26.5
pytesseract==0.3.13
pytest==8.4.2
python-dotenv==1.2.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.49.3
tomli==2.3.0
typing_extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
watchdog==6.0.0
